replaced instruments classes in task1
